Changes based on Comments from Phil Hodge
These are mostly clean up issues. I changed on of the tests for jump at slice boundaries to only run when the number of slices is greater than 1. Otherwise, the test would write outside of the boundaries of the array.

+++ jwst/jump/tests/test_jump_step.py
@@ -80,9 +80,9 @@ def setup_inputs():
         gain = np.ones(shape=(nrows, ncols), dtype=np.float64) * gain
         err = np.ones(shape=(nints, ngroups, nrows, ncols), dtype=np.float64)
         data = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.float64)
-        pixdq = np.zeros(shape=(nrows, ncols), dtype=np.float64)
+        pixdq = np.zeros(shape=(nrows, ncols), dtype=np.uint32)
         read_noise = np.full((nrows, ncols), readnoise, dtype=np.float64)
-        gdq = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.int32)
+        gdq = np.zeros(shape=(nints, ngroups, nrows, ncols), dtype=np.uint32)
 
         rampmodel = MIRIRampModel(data=data, err=err, pixeldq=pixdq, groupdq=gdq, times=times)
         rampmodel.meta.instrument.name = 'MIRI'
